Change DB structure

+++ emo.sh
@@ -1,13 +1,10 @@
 #!/usr/bin/env bash
-# shellcheck disable=SC1090 disable=SC2119
 set -euo pipefail
 
 path_config_dir="$HOME/.config/emo"
 path_data_dir="$HOME/.local/share/emo"
-path_hashes_file="$path_data_dir/hashes"
 path_songs_dir="$path_data_dir/songs"
 path_groups_dir="$path_data_dir/groups"
-path_tags_dir="$path_data_dir/tags"
 
 md() {
 	mkdir -p "$@"
@@ -15,9 +12,8 @@ md() {
 
 # load config
 md "$path_config_dir"
+# shellcheck disable=SC1090
 find "$path_config_dir" -type f | while read -r i; do . "$i"; done
-
-declare -A htn=()
 
 die() {
 	echo "$@" >&2
@@ -26,119 +22,83 @@ die() {
 
 list-something() {
 	path="$1"
-	skip=$((${#path} + 2)) # 2 skips leading /
 	shift
-	find "$path" -mindepth 1 "$@" | cut -c "$skip-"
-}
-
-build-hash-to-name() {
-	((${#htn[@]} > 0)) && return
-	cd "$path_songs_dir"
-	while read -r hash song; do
-		htn["$hash"]="$song"
-	done <<< "$(
-		emo list-songs \
-		| sed -E 's|(.*)|sha256sum "\1"|' \
-		| parallel --will-cite \
-	)"
-	cd "$OLDPWD"
+	(cd "$path"; find . -mindepth 1 "$@") | cut -c 3-
 }
 
 cmd_cleanup() {
 	false && usage ""
-	cmd_list-songs | while read -r i; do
-		cmd_name-to-hash "$i"
-	done > "$path_hashes_file"
 	list-something "$path_data_dir" -type d -empty -delete
 }
 
 cmd_add-song() {
 	(($# < 2)) && usage "<file> <future song name>"
-	md "$path_songs_dir/$(dirname "$2")"
-	cp "$1" "$path_songs_dir/$2"
-	cmd_name-to-hash "$2" >> "$path_hashes_file"
+	md "$path_songs_dir/$2"
+	cp "$1" "$path_songs_dir/file"
 	echo "Added $1 to song library named $2"
 }
 
 cmd_del-song() {
 	(($# < 1)) && usage "<song name>"
-	file="$(mktemp)"
-	grep -v "$(cmd_name-to-hash "$1")" "$path_hashes_file" > "$file"
-	mv "$file" "$path_hashes_file"
-	rm "$path_songs_dir/$1"
+	rm -rf "$path_songs_dir/${1:?}"
 	echo "Removed $1 from song library"
-}
-
-cmd_name-to-hash() {
-	(($# < 1)) && usage "<song name>"
-	sha256sum "$path_songs_dir/$1" | awk '{print $1}'
-}
-
-cmd_hash-to-name() {
-	(($# < 1)) && usage "<hash>"
-	build-hash-to-name
-	set +u
-	name="${htn["$1"]}"
-	[ -z "$name" ] && die "Unknown hash: $1"
-	set -u
-	echo "$name"
 }
 
 cmd_list-songs() {
 	false && usage ""
-	list-something "$path_songs_dir" -type f
+	list-something "$path_songs_dir" -type f -name 'file' | sed 's|/file$||'
 }
 
 cmd_add-to-group() {
 	(($# < 2)) && usage "<song name> <group name>"
-	hash="$(cmd_name-to-hash "$1")"
-	dir="$path_groups_dir/$2"
-	md "$dir"
-	touch "$dir/$hash"
+	md "$path_groups_dir"
+	: >> "$path_groups_dir/$2"
+	sort -u <(echo "$1") -o "$path_groups_dir/$2"{,}
 	echo "Added $1 to group $2"
 }
 
 cmd_del-from-group() {
 	(($# < 2)) && usage "<song name> <group name>"
-	hash="$(cmd_name-to-hash "$1")"
-	rm "$path_groups_dir/$2/$hash"
+	file="$(mktemp)"
+	if grep -Fv "$1" "$path_groups_dir/$2" > "$file"; then
+		mv "$file" "$path_groups_dir/$2"
+	else
+		rm "$path_groups_dir/$2"
+	fi
 	echo "Removed $1 from group $2"
 }
 
 cmd_list-groups() {
 	false && usage ""
-	list-something "$path_groups_dir" -type d
+	ls -1 "$path_groups_dir"
 }
 
 cmd_list-in-group() {
 	(($# < 1)) && usage "<group name>"
-	list-something "$path_groups_dir/$1" -maxdepth 1 -type f | while read -r i; do
-		cmd_hash-to-name "$i"
-	done
+	cat "$path_groups_dir/$1"
 
 }
 
 cmd_set-tag() {
 	(($# < 3)) && usage "<song name> <tag name> <value>"
-	hash="$(cmd_name-to-hash "$1")"
-	dir="$path_tags_dir/$hash"
-	md "$dir"
-	echo "$3" > "$dir/$2"
+	echo "$3" > "$path_songs_dir/$1/$2"
 }
 
 cmd_get-tag() {
-	(($# < 2)) && usage "<song name> <tag name>"
-	cat "$path_tags_dir/$(cmd_name-to-hash "$1")/$2"
+	(($# < 2)) && usage "<song name> <tag names>"
+	song="$1"
+	shift
+	cat "${@/#/"$path_songs_dir/$song/"}"
 }
 
 cmd_del-tag() {
 	(($# < 2)) && usage "<song name> <tag name>"
-	rm "$path_tags_dir/$(cmd_name-to-hash "$1")/$2"
+	rm "$path_songs_dir/$1/$2"
 }
 
 cmd_list-tags() {
 	(($# < 1)) && usage "<song name>"
-	list-something "$path_tags_dir/$(cmd_name-to-hash "$1")"
+	list-something "$path_songs_dir/$1" -type f -not -name "file"
 }
 
 cmd_getcmds() {
